backend/requirements.txt
# BitBot backend — FastAPI + LangGraph + HTTP client to Bento classifier
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
httpx>=0.27.0
langgraph>=0.2.0
psycopg2-binary>=2.9.9

pytest>=8.0.0,<9
pytest-asyncio>=0.24.0
